+++ src/pyinsole/_compat.py
